requirements.txt
openai
requests
PyGithub